Extract <ExerciseTooltip />
