+++ app/javascript/components/student/ExerciseStatusDot.tsx
@@ -1,28 +1,7 @@
-import React, { useState, useEffect, forwardRef } from 'react'
-import { Exercise, ExerciseStatus, Track, SolutionForStudent } from '../types'
-import { Icon, ExerciseWidget } from '../common'
+import React, { forwardRef } from 'react'
+import { ExerciseStatus } from '../types'
 import { usePanel } from '../../hooks/use-panel'
-import { useRequestQuery } from '../../hooks/request-query'
-import { useIsMounted } from 'use-is-mounted'
-import { FetchingBoundary } from '../FetchingBoundary'
-
-const DEFAULT_ERROR = new Error('Unable to load information')
-
-const LoadingComponent = () => {
-  const [isShowing, setIsShowing] = useState(false)
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setIsShowing(true)
-    }, 200)
-
-    return () => clearTimeout(timer)
-  }, [])
-
-  return isShowing ? (
-    <Icon icon="spinner" alt="Loading exercise data" className="--spinner" />
-  ) : null
-}
+import { ExerciseTooltip } from '../tooltips/ExerciseTooltip'
 
 export const ExerciseStatusDot = ({
   slug,
@@ -50,17 +29,6 @@ export const ExerciseStatusDot = ({
     ],
   })
 
-  const isMountedRef = useIsMounted()
-  const { data, error, status } = useRequestQuery<{
-    track: Track
-    exercise: Exercise
-    solution: SolutionForStudent
-  }>(
-    `exercise-tooltip-${slug}`,
-    { endpoint: links.tooltip, options: { enabled: open } },
-    isMountedRef
-  )
-
   return (
     <React.Fragment>
       <ReferenceElement
@@ -71,25 +39,11 @@ export const ExerciseStatusDot = ({
         link={links.exercise}
       />
       {open ? (
-        <div className="c-exercise-tooltip" {...panelAttributes}>
-          <FetchingBoundary
-            status={status}
-            error={error}
-            defaultError={DEFAULT_ERROR}
-            LoadingComponent={LoadingComponent}
-          >
-            {data ? (
-              <ExerciseWidget
-                exercise={data.exercise}
-                solution={data.solution}
-                track={data.track}
-                size="tooltip"
-              />
-            ) : (
-              <span>Unable to load information</span>
-            )}
-          </FetchingBoundary>
-        </div>
+        <ExerciseTooltip
+          slug={slug}
+          endpoint={links.tooltip}
+          {...panelAttributes}
+        />
       ) : null}
     </React.Fragment>
   )
